mem0_memory/pyproject.toml
[project]
name = "mem0-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "attrs==24.3.0",
    "backoff==2.2.1",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "frozenlist==1.5.0",
    "grpcio-tools==1.69.0",
    "grpcio==1.69.0",
    "h11==0.14.0",
    "h2==4.1.0",
    "hpack==4.0.0",
    "httpcore==1.0.7",
    "httpx-sse==0.4.0",
    "httpx==0.28.1",
    "hyperframe==6.0.1",
    "idna==3.10",
    "jiter==0.8.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-community==0.3.14",
    "langchain-core==0.3.29",
    "langchain-text-splitters==0.3.5",
    "langchain==0.3.14",
    "langsmith==0.2.10",
    "litellm>=1.53.3",
    "marshmallow==3.25.1",
    "mem0ai==0.1.44",
    "monotonic==1.6",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "numpy==2.2.1",
    "openai==1.59.7",
    "orjson==3.10.14",
    "packaging==24.2",
    "portalocker==2.10.1",
    "posthog==3.8.3",
    "propcache==0.2.1",
    "protobuf==5.29.3",
    "psycopg2-binary==2.9.10",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "pydantic==2.10.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "qdrant-client==1.12.2",
    "requests-toolbelt==1.0.0",
    "requests==2.32.3",
    "setuptools==75.8.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.37",
    "tenacity==9.0.0",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "urllib3==2.3.0",
    "wheel==0.45.1",
    "yarl==1.18.3",
]
